demos: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@39325 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/demos/font2c.c
@@ -368,7 +368,7 @@ static VOID print_long_array(ULONG *array, ULONG len, FILE *fh)
 	if ((i & 0x03) == 0)
 	    fprintf(fh, "\n\t");
 	    
-    	fprintf(fh,"0x%.8lx, ", *array ++);
+    	fprintf(fh,"0x%.8lx, ", (unsigned long)(*array ++));
 	
     }
     /* print last entry without ',' at the end */
@@ -376,7 +376,7 @@ static VOID print_long_array(ULONG *array, ULONG len, FILE *fh)
     if ((i & 0x03) == 0)
 	fprintf(fh, "\n\t");
 	
-    fprintf(fh, "0x%.8lx\n", *array);
+    fprintf(fh, "0x%.8lx\n", (unsigned long)(*array));
     fprintf(fh, "};");
 
     return;
